+++ SurfaceTopography/IO/LEXT.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2022-2023 Lars Pastewka
+# Copyright 2023-2026 Lars Pastewka
 #
 # ### MIT license
 #
@@ -21,23 +21,51 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 #
-
+#
 # Reference information and implementations:
 # https://sourceforge.net/p/gwyddion/code/HEAD/tree/trunk/gwyddion/modules/file/lextfile.c
+#
 
-import enum
+from ..Exceptions import CorruptFile, FileFormatMismatch
+from .binary import TIFFContainer
+from .expr import C, F, Tup
+from .Reader import Check, CompoundLayout, DeclarativeReaderBase, MagicMatch
 
-import dateutil.parser
-import xmltodict
-from tifffile import TiffFile, TiffFileError
+# The height data lives in the page marked HEIGHT; the measurement
+# metadata is XML embedded in the image description (and the Exif
+# device-setting description) of the overview page
+_height_page = F.match_records(
+    C.pages, ["tags", "ImageDescription"], "HEIGHT"
+)[0]
+_metadata_page = F.match_records_containing(
+    C.pages, ["tags", "ImageDescription"], "TiffTagDescData"
+)[0]
 
-from ..Exceptions import CorruptFile, FileFormatMismatch, MetadataAlreadyFixedByFile
-from ..UniformLineScanAndTopography import Topography
-from .common import OpenFromAny
-from .Reader import ChannelInfo, ReaderBase
+_image_desc = F.parse_xml(_metadata_page.tags.ImageDescription)[
+    "TiffTagDescData"
+]
+_data_desc = F.parse_xml(
+    _metadata_page.tags.ExifTag.DeviceSettingDescription
+)["ExifTagDescData"]
+
+_height_info = _image_desc.HeightInfo
+_calibration = _data_desc.ImageCommonSettingsInfo
+
+_nb_grid_pts_x = _metadata_page.shape[0]
+_nb_grid_pts_y = _metadata_page.shape[1]
 
 
-class LEXTReader(ReaderBase):
+def _physical_size(nb_grid_pts, data_per_pixel, calibration_value):
+    # All lengths appear to be picometers; we report micrometers
+    return (
+        1e-12
+        * nb_grid_pts
+        * F.float(_height_info[data_per_pixel])
+        * F.float(_calibration[calibration_value])
+    )
+
+
+class LEXTReader(DeclarativeReaderBase):
     _format = "lext"
     _mime_types = ["image/tiff"]
     _file_extensions = ["lext"]
@@ -47,198 +75,92 @@ class LEXTReader(ReaderBase):
 TIFF-based file format of Olympus instruments.
 """
 
-    _tag_readers = {306: lambda value: dateutil.parser.parse(value)}
+    # TIFF magic bytes (little- and big-endian)
+    _MAGIC_TIFF = (b"II*\x00", b"MM\x00*")
+
+    # A TIFF magic can only tell that the file *may* be a LEXT file;
+    # detection needs to try parsing
+    _magic = None
 
     @classmethod
-    def _tag_to_dict(cls, key, value):
-        # Try dedicated tag readers first
-        try:
-            return {value.name: cls._tag_readers[key](value.value)}
-        except KeyError:
-            pass
+    def can_read(cls, buffer: bytes) -> MagicMatch:
+        if len(buffer) < 4:
+            return MagicMatch.MAYBE  # Buffer too short to determine
+        if buffer.startswith(cls._MAGIC_TIFF):
+            # TIFF file, could be an Olympus LEXT file
+            return MagicMatch.MAYBE
+        return MagicMatch.NO
 
-        # Generic treatment if failed
-        if isinstance(value.value, enum.Enum):
-            return {value.name: [value.value.name, value.value.value]}
-        else:
-            name = value.name
-            value = value.value
-            try:
-                if value.startswith(r"<?xml"):
-                    value = xmltodict.parse(value)
-            except AttributeError:
-                pass
-            if isinstance(value, tuple):
-                value = list(value)
-            return {name: value}
-
-    # Reads in the positions of all the data and metadata
-    def __init__(self, file_path):
-        self._file_path = file_path
-
-        # All units appear to be picometers, also some files report 'INCH' as resolution unit
-        self._unit = "µm"
-
-        with OpenFromAny(self._file_path, "rb") as f:
-            try:
-                with TiffFile(f) as t:
-                    # Go through all pages and see what is in there
-                    metadata = []
-                    for i, p in enumerate(t.pages):
-                        page_metadata = {}
-                        for key, value in p.tags.items():
-                            tag_metadata = self._tag_to_dict(key, value)
-                            if "ExifTag" in tag_metadata:
-                                updated_metadata = {}
-                                for key, value in tag_metadata["ExifTag"].items():
-                                    try:
-                                        if value.startswith("<?xml"):
-                                            updated_metadata[key] = xmltodict.parse(
-                                                value
-                                            )
-                                    except AttributeError:
-                                        pass
-                                tag_metadata["ExifTag"].update(updated_metadata)
-                            page_metadata.update(tag_metadata)
-
-                        if "DateTime" in page_metadata:
-                            page_metadata["DateTime"] = page_metadata[
-                                "DateTime"
-                            ].isoformat()
-
-                        if page_metadata["ImageDescription"] == "HEIGHT":
-                            # This is the actual height data
-                            self._height_page_index = i
-                        elif "TiffTagDescData" in page_metadata["ImageDescription"]:
-                            # Get image dimension
-                            try:
-                                nx, ny, _ = p.shape
-                            except ValueError:
-                                nx, ny = p.shape
-
-                            # This is some metadata
-                            image_desc = page_metadata["ImageDescription"][
-                                "TiffTagDescData"
-                            ]
-                            data_desc = page_metadata["ExifTag"][
-                                "DeviceSettingDescription"
-                            ]["ExifTagDescData"]
-                            self._height_scale_factor = (
-                                1e-12
-                                * float(image_desc["HeightInfo"]["HeightDataPerPixelZ"])
-                                * float(
-                                    data_desc["ImageCommonSettingsInfo"][
-                                        "MakerCalibrationValueZ"
-                                    ]
-                                )
-                            )
-                            self._nb_grid_pts = (nx, ny)
-                            self._physical_sizes = (
-                                1e-12
-                                * nx
-                                * float(image_desc["HeightInfo"]["HeightDataPerPixelX"])
-                                * float(
-                                    data_desc["ImageCommonSettingsInfo"][
-                                        "MakerCalibrationValueX"
-                                    ]
-                                ),
-                                1e-12
-                                * ny
-                                * float(image_desc["HeightInfo"]["HeightDataPerPixelY"])
-                                * float(
-                                    data_desc["ImageCommonSettingsInfo"][
-                                        "MakerCalibrationValueY"
-                                    ]
-                                ),
-                            )
-                            self._x_unit_index = int(
-                                image_desc["HeightInfo"]["HeightDataUnitX"]
-                            )
-                            self._y_unit_index = int(
-                                image_desc["HeightInfo"]["HeightDataUnitY"]
-                            )
-                            self._height_unit_index = int(
-                                image_desc["HeightInfo"]["HeightDataUnitZ"]
-                            )
-
-                            if (
-                                self._x_unit_index != self._height_unit_index
-                                or self._y_unit_index != self._height_unit_index
-                            ):
-                                raise CorruptFile(
-                                    "Units in x-, y- and height direction appear to "
-                                    "differ."
-                                )
-
-                        metadata += [page_metadata]
-
-                    self._info = {"raw_metadata": metadata}
-            except TiffFileError:
-                raise FileFormatMismatch(
-                    "This is not a TIFF file, so it cannot be an Olympus LEXT file."
+    _file_layout = CompoundLayout(
+        [
+            TIFFContainer(),
+            Check(
+                F.len(
+                    F.match_records(
+                        C.pages, ["tags", "ImageDescription"], "HEIGHT"
+                    )
                 )
-
-    @property
-    def channels(self):
-        return [
-            ChannelInfo(
-                self,
-                0,  # channel index
-                name="default",
-                dim=2,
-                nb_grid_pts=self._nb_grid_pts,
-                physical_sizes=self._physical_sizes,
-                height_scale_factor=self._height_scale_factor,
-                uniform=True,
-                info=self._info,
-                unit=self._unit,
-            )
+                == 1,
+                FileFormatMismatch,
+                "TIFF file does not contain a HEIGHT page, so this is not "
+                "an Olympus LEXT file.",
+            ),
+            Check(
+                F.len(
+                    F.match_records_containing(
+                        C.pages, ["tags", "ImageDescription"],
+                        "TiffTagDescData",
+                    )
+                )
+                == 1,
+                FileFormatMismatch,
+                "TIFF file does not contain LEXT measurement metadata, so "
+                "this is not an Olympus LEXT file.",
+            ),
+            Check(
+                (
+                    F.int(_height_info.HeightDataUnitX)
+                    == F.int(_height_info.HeightDataUnitZ)
+                )
+                & (
+                    F.int(_height_info.HeightDataUnitY)
+                    == F.int(_height_info.HeightDataUnitZ)
+                ),
+                CorruptFile,
+                "Units in x-, y- and height direction appear to differ.",
+            ),
         ]
+    )
 
-    def topography(
-        self,
-        channel_index=None,
-        physical_sizes=None,
-        height_scale_factor=None,
-        unit=None,
-        info={},
-        periodic=None,
-        subdomain_locations=None,
-        nb_subdomain_grid_pts=None,
-    ):
-        if subdomain_locations is not None or nb_subdomain_grid_pts is not None:
-            raise RuntimeError("This reader does not support MPI parallelization.")
-
-        if channel_index is None:
-            channel_index = self._default_channel_index
-
-        if channel_index != self._default_channel_index:
-            raise RuntimeError(
-                f"There is only a single channel. Channel index must be {self._default_channel_index}."
-            )
-
-        if physical_sizes is not None:
-            raise MetadataAlreadyFixedByFile("physical_sizes")
-
-        if height_scale_factor is not None:
-            raise MetadataAlreadyFixedByFile("height_scale_factor")
-
-        if unit is not None:
-            raise MetadataAlreadyFixedByFile("unit")
-
-        with OpenFromAny(self._file_path, "rb") as f:
-            with TiffFile(f) as t:
-                height_data = t.pages[self._height_page_index].asarray()
-                assert height_data.shape == self._nb_grid_pts
-
-        _info = self._info.copy()
-        _info.update(info)
-
-        topo = Topography(
-            height_data,
-            self._physical_sizes,
-            unit=self._unit,
-            periodic=False if periodic is None else periodic,
-            info=_info,
-        )
-        return topo.scale(self._height_scale_factor)
+    _channel_bindings = [
+        {
+            "name": "default",
+            "dim": 2,
+            "nb_grid_pts": Tup(_nb_grid_pts_x, _nb_grid_pts_y),
+            "physical_sizes": Tup(
+                _physical_size(
+                    _nb_grid_pts_x,
+                    "HeightDataPerPixelX",
+                    "MakerCalibrationValueX",
+                ),
+                _physical_size(
+                    _nb_grid_pts_y,
+                    "HeightDataPerPixelY",
+                    "MakerCalibrationValueY",
+                ),
+            ),
+            "height_scale_factor": 1e-12
+            * F.float(_height_info.HeightDataPerPixelZ)
+            * F.float(_calibration.MakerCalibrationValueZ),
+            "uniform": True,
+            "unit": "µm",
+            "info": {
+                "raw_metadata": {
+                    "pages": F.pluck(C.pages, "tags"),
+                    "measurement": _image_desc,
+                    "device_settings": _data_desc,
+                },
+            },
+            "data": _height_page.image,
+        }
+    ]